+++ CloudHarvestApi/cache/exceptions.py
@@ -1,6 +1,0 @@
-from CloudHarvestCoreTasks.exceptions import BaseHarvestException
-
-
-class HarvestCacheException(BaseHarvestException):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)